+++ website/CHANGELOG.md
@@ -6,6 +6,7 @@ release.
 ## 0.2.1
 
 * New `Plugs: Add` command
+* When holding `Shift` while pasting, rich test paste will be disabled.
 
 ## 0.2.0
 * The editor is now in "live preview" mode where a lot of markdown is hidden unless the cursor is present. This will take some getting used to, but results in a much more distraction free look.